Move OrtError into core crate
I need it in there so that core can use it (`core` can't depend on `cli`,
obviously).

Unforunately that means I can't implement `From<std::io::Error>` for
`OrtError`. The impl needs to either be in the same crate as
std::io::Error (no can do) or core would have to import that type, which
is std not core so we also can't do that.

Which means `?` can't auto-convert (by calling `Into` on the error
type), hence the profusion of `.map_err(ort_from_err)`. Oh well.