Use pthread instead of direct clone for thread creation
pthread threads and Rust (which uses pthread internally) rely on there
being a valid thread-local storage. Rust's stdout lock for example
depends on it to work correctly.

To use `clone` directly I would need to create a pthread compatible
thread-local storage area. So let pthread do it, this is what Rust does.

Credit to gpt-5.1-codex-max for figuring it out.